setcolors.vim: Experiment with gruvbox, jellybeans, and hybrid
The setcolors.vim script is sourced very late in the list of scriptnames,
so the SetColors command is not available while sourcing the vimrc file.
SetColors in a VimEnter autocmd instead, so that it only happens once
the secolors.vim script has been sourced.

Each time we rotate to another colorscheme with <F8>,
my preferred transparent background gets turned back off,
so turn it back on after rotating the colorscheme.
This also hides the name of the new colorscheme,
so we display that again as well.